Yolo11/requirements.txt
streamlit
ultralytics
opencv-python-headless
Pillow
streamlit-webrtc